projects/minpaku-blog/requirements.txt
requests>=2.32
python-dotenv>=1.0
gspread>=6.0
google-auth>=2.30
